// src/project/constants/paths.enum.ts
export enum Paths {
    PROJECT = "project",
    PARAM_PROJECT = ":projectId",
    PARAM_PROJECT_ID = "projectId",

    NODE = "node",
    PARAM_NODE = ":nodeId",
    PARAM_NODE_ID = "nodeId",

    COLLABORATOR = "collaborator",
    PARAM_COLLABORATOR = ":collaboratorId",
    PARAM_COLLABORATOR_ID = "collaboratorId",

    
    BY_TOKEN = "byToken",
    BY_TOKEN_PROJECT = "byTokenProject",
    BY_TOKEN_CLIENT = "byTokenClient"
}